Finished the html portion of butter chicken page
I used an unordered list to display the ingredients needed for the
recipe, and used an ordered list to dispay the recipe instructions.

Format of the page is the same as the grilled cheese page in which
its title, image, description, ingredients and directions.

I think latter down the line we could add a section to write and paste
reviews for the page, to help increase user interactivity with the page.
Although for now I think it would be better to stile the page first.