Completed the pancake recipe page
Page is not ready for css, later on. Format of the page follows is the
same as the grilled cheese page, and the butter chicken page.

I think changes that could be made later are positiong the elements,
such as making the title in the center and having ingredients and
directions be side by side.